+++ frontend/src/pages/CatalogPage/SectionContent/SectionContent.tsx
@@ -12,39 +12,20 @@ interface SectionContentProps {
 }
 
 const SectionContent: React.FC<SectionContentProps> = () => {
-  const [searchParams, setSearchParams] = useSearchParams();
-  const { currentPage, setCurrentPage, totalPages } = useBooks();
-
-  const handlePageChange = (page: number) => {
-    if (page !== currentPage) {
-      setCurrentPage(page);
-    }
-    const keyword = searchParams.get("keyword") || "";
-
-    setSearchParams({
-      keyword,
-      page: page.toString(),
-    });
-  };
   const { books } = useOutletContext<{ books: IBookItem[] }>();
 
   const hasBooks = hasData(books);
 
   return (
-    <StyledFlexWrap>
-      {hasBooks ? (
-        books.map((book, index) => <CatalogBookCard key={index} {...book} />)
-      ) : (
-        <TextCenter>No books in catalog</TextCenter>
-      )}
-      {totalPages > 1 && (
-        <Pagination
-          currentPage={currentPage}
-          totalPages={totalPages}
-          onPageChange={handlePageChange}
-        />
-      )}
-    </StyledFlexWrap>
+    <>
+      <StyledFlexWrap>
+        {hasBooks ? (
+          books.map((book, index) => <CatalogBookCard key={index} {...book} />)
+        ) : (
+          <TextCenter>No books in catalog</TextCenter>
+        )}
+      </StyledFlexWrap>
+    </>
   );
 };
